print sizes

+++ 0x00-hello_world/6-size.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 /**
-* main - Entry point
+* ma in - Entry point
 *Return: Always 0 (success)
 */
 int main(void)
 {
-	printf("Size of a char: %ld byte(S)\n", sizeof(char));
-	printf("Size of an int: %ld byte(S)\n", sizeof(int));
-	printf("Size of a long int: %ld byte(S)\n", sizeof(long int));
+	printf("Size of a char: %lu byte(S)\n", sizeof(char));
+	printf("Size of an int: %ld byte(S)\n", sizeof(long unsigned int));
+	printf("Size of a long int: %ld byte(S)\n", sizeof(long unsigned  int));
 	printf("Size of a long long int: %ld byte(S)\n", sizeof(long long int));
 	printf("Size of a float: %ld byte(S)\n", sizeof(long unsigned int));
 		return (0);
